# Conflicts:
#	android/app/src/main/java/com/ssafy/keywe/presentation/auth/LoginScreen.kt
#	android/app/src/main/java/com/ssafy/keywe/util/NetworkUtil.kt

+++ android/app/src/main/java/com/ssafy/keywe/common/KeyWeNavigation.kt
@@ -44,7 +44,7 @@ sealed interface Route {
         data object MenuRoute : Route
 
         @Serializable
-        data class MenuDetailRoute(val id: Int) : Route
+        data class MenuDetailRoute(val id: Long) : Route
 
         @Serializable
         data object MenuCartRoute : Route
